requirements.txt
# oTree-may-overwrite-this-file
# IF YOU MODIFY THIS FILE, remove these comments. 
# otherwise, oTree will automatically overwrite it.

# Core oTree Dependencies
otree>=5.10.0

# Configuration Management
PyYAML>=6.0

# Database Support
psycopg2-binary>=2.9.0

# Error Monitoring (Optional but recommended for production)
sentry-sdk>=1.0.0

# Optional: Data Analysis & Visualization (useful for experiment analysis)
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Optional: Advanced Features
# Uncomment if you need these features:
# matplotlib>=3.5.0  # For data visualization
# seaborn>=0.11.0    # For statistical plotting
# jupyter>=1.0.0     # For interactive analysis

# Development & Deployment
# Uncomment for production deployment:
# gunicorn>=20.1.0   # WSGI server for production
# uvicorn>=0.15.0    # ASGI server for WebSocket support
